Refactor: 커스텀 훅 생성 및 리펙토링
